% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_database}
\alias{download_database}
\title{RExMap database from RExMapDB}
\usage{
download_database(delete_old = TRUE, verbose = TRUE, use_api = FALSE)
}
\arguments{
\item{delete_old}{Delete old files during an update.}

\item{verbose}{Display progress messeges.}

\item{use_api}{Use GitHub API. Warning: has 60 requests/hr hard limit.}
}
\description{
Update/replace database with RExMapDB latest version. This deletes the
current database files from find.package('rexmap')/inst/database, then
downloads the files from RExMapDB Github repository.
}
